requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
passlib
python-jose